+++ server/src/main/kotlin/ru/spbstu/application/auth/repository/UserRepository.kt
@@ -10,6 +10,6 @@ interface UserRepository {
     fun contains(phoneNumber: PhoneNumber): Boolean
     fun contains(id: User.Id): Boolean
     fun sortByAmountOfCoins(): List<User>
-    fun setAmountOfCoins(id: User.Id, newAmountOfCoins: Long)
-    fun setAvailableStepsCount(id: User.Id, newAvailableStepsCount: Long)
+    fun setAmountOfCoins(id: User.Id, newAmountOfCoins: Int)
+    fun setAvailableStepsCount(id: User.Id, newAvailableStepsCount: Int)
 }
