+++ server/src/main/kotlin/ru/spbstu/application/steps/entities/Step.kt
@@ -4,12 +4,12 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 @JvmInline
-value class Step(val value: Long) {
+value class Step(val value: Int) {
     init {
         require(value in 0..LastValue)
     }
 
     companion object {
-        const val LastValue = 4L
+        const val LastValue = 4
     }
 }
